requirements.txt
praw
google-generativeai
google-api-python-client